Posztok törlésének kiegészítése. Mostantól a posztok törlése mellett a képeket is törli a rendszer a bucket-ből, így nem hagy maga után semmit a felhasználó.

+++ Angular project/SocialMedia/src/app/Shared/services/post.service.ts
@@ -1,6 +1,8 @@
 import { Injectable } from '@angular/core';
 import {AngularFirestore} from "@angular/fire/compat/firestore";
 import {Posztok} from "../../Model/posztok";
+import {FileUploadService} from "./file-upload.service";
+import {finalize, first} from "rxjs";
 
 @Injectable({
   providedIn: 'root'
@@ -9,7 +11,7 @@ export class PostService {
 
   postCollectionName = "Posztok";
 
-  constructor(private afs: AngularFirestore) { }
+  constructor(private afs: AngularFirestore, private fileUploadService: FileUploadService) { }
 
   create(poszt: Posztok) {
     return this.afs.collection(this.postCollectionName).doc(poszt.postID).set(poszt);
@@ -37,7 +39,15 @@ export class PostService {
     // await postRef.update({
     //   postCount: FieldValue.increment(-1)
     // });
-    return this.afs.collection(this.postCollectionName).doc(postId).delete();
+    let kepid;
+    this.getPoszt(postId).valueChanges().pipe(first()).subscribe(post => {
+      kepid = post.kepId;
+      if (kepid === "") return this.afs.collection(this.postCollectionName).doc(postId).delete();
+      this.fileUploadService.delete(kepid).then(_ => {
+        console.log(kepid + "  lefut?")
+        return this.afs.collection(this.postCollectionName).doc(postId).delete();
+      });
+    });
   }
 
   //Balazs belenyul, hogy tonkretegye a kodot, Puszi!
